this is histogram stuff as well as data outputs from catsrc

+++ source_handling/clus_subtract_xcomps.py
@@ -24,11 +24,12 @@ from clus_get_data import *
 from astropy.convolution import convolve_fft
 # from scipy import optimize.least_squares
 from scipy.stats import linregress
-from writefits import *
+from writefits import writefits
 from gaussian import makeGaussian, padGaussian
 from clus_make_noise_mask import clus_make_noise_mask
+import config
 
-def clus_subtract_xcomps(maps, simflag=0, verbose=1, superplot=1):
+def clus_subtract_xcomps(maps, simflag=0, verbose=1, superplot=1, nsim=0, saveplot=1):
 
     err =  None
     ncols = len(maps)
@@ -76,13 +77,20 @@ def clus_subtract_xcomps(maps, simflag=0, verbose=1, superplot=1):
         y = slope * PSW_array + intercept
 
 
-        if superplot:
+        if superplot or saveplot:
             plt.plot(PSW_array, ref_array, 'x')
             plt.plot(PSW_array, y, c='red')
             plt.title('clus_subtract_xcomps: PSW vs. %s' %(maps[i]['band']))
             plt.xlabel('PSW')
             plt.ylabel('%s' %(maps[i]['band']))
-            plt.show()
+            if superplot:
+                plt.show()
+            elif saveplot:
+                if nsim != 0:
+                    filename = config.HOME + 'outputs/correlated_components/' + maps[i]['clusname'] + '_' + maps[i]['band'] + '_' + str(nsim) + '.pdf'
+                else:
+                    filename = config.HOME + 'outputs/correlated_components/' + maps[i]['clusname'] + '_' + maps[i]['band'] + '_' + str(nsim) + '.pdf'
+                plt.savefig(filename, format='pdf')
 
         #subtract the correlated components from the image
         maps[i]['xclean'] = np.empty(maps[i]['srcrm'].shape)
@@ -100,15 +108,16 @@ def clus_subtract_xcomps(maps, simflag=0, verbose=1, superplot=1):
             plt.title('xclean map')
             plt.show()
 
-        if not simflag:
-            filename = config.CLUSDATA + 'sz/' + maps[i]['name'] + str(maps[i]['band']) + '_xc.fits'
-        else:
-            filename = config.CLUSDATA + 'sz/sim/' + maps[i]['name'] + str(maps[i]['band']) + '_xc.fits'
+        if saveplot:
+            if not simflag:
+                filename = config.HOME + 'outputs/correlated_components/' + maps[i]['name'] + '_' + maps[i]['band'] + '_xc.fits'
+            else:
+                filename = config.HOME + 'outputs/correlated_components/' + maps[i]['name'] + '_' + maps[i]['band'] + '_' + str(nsim) + '_xc.fits'
 
-        filename = 'correlated_comp_test_%s.fits' % (maps[i]['band'])
-        if os.path.isfile(filename):
-            os.remove(filename)
-        writefits(filename, data=datasub, header_dict=maps[i]['shead'])
+            filename = 'correlated_comp_test_%s.fits' % (maps[i]['band'])
+            if os.path.isfile(filename):
+                os.remove(filename)
+            writefits(filename, data=datasub, header_dict=maps[i]['shead'])
 
     #subtract the mean of the new map from itself.
     # for i in range(maps[0]['xclean'].shape[0]):
